update setting fragment
