+++ core/nav/src/main/java/com/kotlin/learn/core/nav/navigator/MenuNavigator.kt
@@ -18,4 +18,10 @@ class MenuNavigator {
         )
     }
 
+    fun fromMenuToSetting(activity: FragmentActivity) {
+        initNavHostFragment(activity).findNavController().navigate(
+            MenuFragmentDirections.actionMenuFragmentToSettingFragment()
+        )
+    }
+
 }
