+++ feature/movie/src/main/java/com/kotlin/learn/feature/movie/presentation/ui/SeeAllCreditFragment.kt
@@ -1,23 +1,12 @@
 package com.kotlin.learn.feature.movie.presentation.ui
 
-import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.navArgs
-import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
 import com.kotlin.learn.core.common.base.BaseFragment
-import com.kotlin.learn.core.common.util.network.Result
-import com.kotlin.learn.core.model.CastModel
-import com.kotlin.learn.core.model.CreditsModel
-import com.kotlin.learn.core.model.CrewModel
 import com.kotlin.learn.core.utilities.Constant
-import com.kotlin.learn.core.utilities.extension.launch
 import com.kotlin.learn.feature.movie.adapter.DetailCreditsAdapter
-import com.kotlin.learn.feature.movie.adapter.SeeAllCreditsAdapter
 import com.kotlin.learn.feature.movie.databinding.FragmentSeeAllCreditBinding
-import com.kotlin.learn.feature.movie.databinding.SeeAllCreditsListBinding
-import com.kotlin.learn.feature.movie.presentation.viewmodel.CreditsViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -39,14 +28,14 @@ class SeeAllCreditFragment : BaseFragment<FragmentSeeAllCreditBinding>(FragmentS
     private fun setupVpCredits() = with(binding) {
         val creditsAdapter = DetailCreditsAdapter(activity)
         creditsAdapter.addFragment(
-            fragment = SeeAllCreditsFragmentView(
+            fragment = SeeAllCreditsViewFragment(
                 movieId = movieId,
                 creditsCategory = CreditsCategory.CAST,
             ),
             title = "Cast"
         )
         creditsAdapter.addFragment(
-            fragment = SeeAllCreditsFragmentView(
+            fragment = SeeAllCreditsViewFragment(
                 movieId = movieId,
                 creditsCategory = CreditsCategory.CREW,
             ),
@@ -64,87 +53,6 @@ class SeeAllCreditFragment : BaseFragment<FragmentSeeAllCreditBinding>(FragmentS
         TabLayoutMediator(tlSeeAllCredits, vpSeeAllCredits) { tab, position ->
             tab.text = creditsAdapter.getTabTitle(position)
         }.attach()
-    }
-
-}
-
-@AndroidEntryPoint
-class SeeAllCreditsFragmentView(
-    private val movieId: String,
-    private val creditsCategory: CreditsCategory,
-) : BaseFragment<SeeAllCreditsListBinding>(SeeAllCreditsListBinding::inflate) {
-
-    private val castAdapter = SeeAllCreditsAdapter.Cast(this::onClickCreditsCast)
-    private val crewAdapter = SeeAllCreditsAdapter.Crew(this::onClickCreditsCrew)
-
-    private val viewModel: CreditsViewModel by viewModels()
-
-    override fun setupView() {
-        subscribeCredits()
-        setupAdapter()
-        setupInit()
-    }
-
-    private fun subscribeCredits() = with(binding) {
-        viewModel.creditsMovie.launch(this@SeeAllCreditsFragmentView) {
-            when (it) {
-                is Result.Waiting -> {}
-
-                is Result.Loading -> {
-                    viewAnimator.displayedChild = 0
-                }
-
-                is Result.Success -> {
-                    viewAnimator.displayedChild = 1
-                    loadContent(it.data)
-                }
-
-                is Result.Error -> {
-                    viewAnimator.displayedChild = 2
-                    showErrorView(
-                        it.throwable.message
-                            ?: "Error occurred when fetching data from server. Please try again"
-                    )
-                }
-            }
-        }
-    }
-
-    private fun loadContent(it: CreditsModel) {
-        if (checkCreditsIsCrew(creditsCategory))
-            crewAdapter.submitList(it.crew)
-        else
-            castAdapter.submitList(it.cast)
-    }
-
-    private fun setupInit() {
-        viewModel.getCredits(movieId)
-    }
-
-    private fun setupAdapter() = with(binding) {
-        rvSeeAllCredits.apply {
-            layoutManager = GridLayoutManager(requireContext(), 3, RecyclerView.VERTICAL, false)
-            adapter = if (checkCreditsIsCrew(creditsCategory)) crewAdapter else castAdapter
-        }
-    }
-
-    private fun showErrorView(message: String) {
-        binding.viewCommonError.apply {
-            errorMessage.text = message
-            buttonRetry.setOnClickListener { setupInit() }
-        }
-    }
-
-    private fun checkCreditsIsCrew(creditsCategory: CreditsCategory): Boolean {
-        return creditsCategory.name == CreditsCategory.CREW.name
-    }
-
-    private fun onClickCreditsCrew(item: CrewModel) {
-
-    }
-
-    private fun onClickCreditsCast(item: CastModel) {
-
     }
 
 }
